Refactor count types & how metadata rows counted
This moves the CountType contructors from the extract module over to the
main file of the library, closer to the type they implement on. The
number of metadata rows in a file is now determined dynamically, rather
than being set by a hard-coded value. Additionally, the constructor
for CountType based on the header is improved and simplified,
by searching for matching header text throughtout the file and without
the previous use of CountType from location.

Tests were also added (which led to the refactoring in the first place).