Calculate non-normal speed averages
These are for another non-normalized table, TC_SPESUM.

Additionally, use lane in the keys for both this and NonNormalVolCount,
which enables getting total for each direction.

Fix various tests as a result. Test that speed averages are calculated
correctly (against manual calculations of raw data). Test that hours
with incomplete data (start/end of count) are not included.