Use discriminants in VehicleClass, handle num 15
Discrimants seem to be slightly clearer than using comments, and in
future could be used in the `from_num` method, though nothing in std
to do that now.

Also, because 15 is supposed to be the unclassified class, treat it the
same as 14 (which is supposed to be "unused" but JAMAR/Starnext uses
it for unclassified).