Generalize non-normal agg, include all data
Non-normal aggregation had previously been tightly coupled with how
class counts are done, but by creating a more general struct for the
data and relying on SQL to do much of the work, this can be applied
for multiple kinds. It's also better to get use the database for this,
since we may want to redo it on old counts we may not have original
data for.

Additionally, periods with incomplete data were previously excluded
from being inserted/aggregrated, but given backwards compatibility
issues and possibly inconsistent methods for exporting data from
counters, they are now included and a note to that effect has been
added to main documentation. Many tests had to be updated as a result
(and also found a bug caused by non sorting vecs in process).