Refactor into multiple templates; improve html
This eliminates the AdminAction enum, which wasn't a perfect fit for
handling because it was being split between the controller and the view.
Now, there are separate routes and templates to handle the various
actions an admin can take. One benefit aside from reducing complexity
is that two dependencies were also removed.

Also further improvements to semantic html.