Add new types of directions
Previously, Direction was just used for extracting data from files, and
there the direction was just the four directions. However, for the
Metadata records, "both" is a valid direction. So, the old Direction
became LaneDirection, and a new one, RoadDirection, was created for
Metadata. Additionally, because both std's Option and oracle's FromSql
are external, I had to create a newtypes, OptionLaneDirection and
OptionRoadDirection, in order to impl FromSql on the Option on these,
since they can be null int the database.