Allow creation of new count(s) from existing one
There's also a lot of other changes that were precipitated by this -
the largest being how messages are passed to the user. Because I needed
to redirect from one route to another, I couldn't use the message
fields in the templates, so I created a global Mutex that handles this,
and in all it turned out to be a more elegant way to handle messages.

I also removed display_some in favor of a custom filter, which doesn't
affect functionality but greatly reduces verbosity.

Finally, some fairly minor stuff, which came along for the ride with
the rest of the changes: improved documentation and a bug fix with
pagination not getting the first page of results.