Use new field to identify one-way bicycle counts
The previous way could cause issue with overcounting because sometimes
the count from a single counting machine is split in two and one dir
used for each recordnum in our database. In these cases, even though
cldir2 is null, we don't want to use the total column in the csv from
ecocounter. The new() fn on Directions was removed so that manual
struct creation has to be used to make setting one_way_bicycle field
more transparent.

Also improve documentation here and there.